`cd benchmarks/handle-event && npm run start` in my environment with this PR.

```
hono - trie-router(default) x 102,452 ops/sec ±5.76% (85 runs sampled)
hono - regexp-router x 145,484 ops/sec ±2.92% (85 runs sampled)
itty-router x 72,637 ops/sec ±3.09% (85 runs sampled)
sunder x 134,340 ops/sec ±2.42% (87 runs sampled)
worktop x 81,620 ops/sec ±4.09% (84 runs sampled)
Fastest is hono - regexp-router
```